Divided all classes and function to different modules
