+++ index.html
@@ -13,9 +13,9 @@
     <div class="topNav flex">
       <h1 href="#" class="logo">Awesome Books</h1>
       <ul class="navbar flex">
-        <li><a href="#" class="nav-link1 active">List</a></li>
-        <li><a href="#" class="nav-link2">Add new</a></li>
-        <li><a href="#" class="nav-link3">Contact</a></li>
+        <li><a href="#" class="nav-link">List</a></li>
+        <li><a href="#" class="nav-link">Add new</a></li>
+        <li><a href="#" class="nav-link">Contact</a></li>
       </ul>
     </div>
     <div class="bottomNav flex">
@@ -23,7 +23,7 @@
     </div>
   </header>
 
-  <section id="awesome-books" class="container flex column">
+  <section id="awesome-books" class="container flex column invisible">
     <h2>All awesome books</h2>
     <ul class="books flex column"></ul>
     <div class="decoration"></div>
