+++ modules/classes/node.mjs
@@ -1,0 +1,6 @@
+export default class Node {
+  constructor(value) {
+    this.value = value;
+    this.next = null;
+  }
+}
